+++ modules/App/GraphQL/Types/JsonType.php
@@ -13,8 +13,8 @@ use GraphQL\Type\Definition\ScalarType;
 
 class JsonType extends ScalarType
 {
-    public $name = 'JsonType';
-    public $description =
+    public string $name = 'JsonType';
+    public ?string $description =
         'The `JSON` scalar type represents JSON values as specified by
         [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf).';